move primitive properties to bokeh.core.property.primitive
